Consolidate factory package into one single file: note that no import cycle change
