+++ validator/accounts/BUILD.bazel
@@ -41,8 +41,6 @@ go_library(
         "//validator/client:go_default_library",
         "//validator/client/beacon-api:go_default_library",
         "//validator/client/iface:go_default_library",
-        "//validator/client/node-client-factory:go_default_library",
-        "//validator/client/validator-client-factory:go_default_library",
         "//validator/helpers:go_default_library",
         "//validator/keymanager:go_default_library",
         "//validator/keymanager/derived:go_default_library",
