+++ validator/client/BUILD.bazel
@@ -8,6 +8,7 @@ go_library(
         "attest.go",
         "duties.go",
         "duty_store.go",
+        "factory.go",
         "health_monitor.go",
         "key_reload.go",
         "log.go",
@@ -65,10 +66,9 @@ go_library(
         "//time/slots:go_default_library",
         "//validator/accounts/iface:go_default_library",
         "//validator/accounts/wallet:go_default_library",
-        "//validator/client/beacon-chain-client-factory:go_default_library",
+        "//validator/client/beacon-api:go_default_library",
+        "//validator/client/grpc-api:go_default_library",
         "//validator/client/iface:go_default_library",
-        "//validator/client/node-client-factory:go_default_library",
-        "//validator/client/validator-client-factory:go_default_library",
         "//validator/db:go_default_library",
         "//validator/db/common:go_default_library",
         "//validator/graffiti:go_default_library",
