--- validator/client/beacon-chain-client-factory/BUILD.bazel
@@ -1,16 +1,0 @@
-load("@prysm//tools/go:def.bzl", "go_library")
-
-go_library(
-    name = "go_default_library",
-    srcs = ["beacon_chain_client_factory.go"],
-    importpath = "github.com/OffchainLabs/prysm/v7/validator/client/beacon-chain-client-factory",
-    visibility = ["//visibility:public"],
-    deps = [
-        "//config/features:go_default_library",
-        "//validator/client/beacon-api:go_default_library",
-        "//validator/client/grpc-api:go_default_library",
-        "//validator/client/iface:go_default_library",
-        "//validator/client/node-client-factory:go_default_library",
-        "//validator/helpers:go_default_library",
-    ],
-)
