+++ validator/client/factory.go
@@ -1,13 +1,20 @@
-package beacon_chain_client_factory
+package client
 
 import (
 	"github.com/OffchainLabs/prysm/v7/config/features"
 	beaconApi "github.com/OffchainLabs/prysm/v7/validator/client/beacon-api"
 	grpcApi "github.com/OffchainLabs/prysm/v7/validator/client/grpc-api"
 	"github.com/OffchainLabs/prysm/v7/validator/client/iface"
-	nodeClientFactory "github.com/OffchainLabs/prysm/v7/validator/client/node-client-factory"
 	validatorHelpers "github.com/OffchainLabs/prysm/v7/validator/helpers"
 )
+
+func NewNodeClient(validatorConn *validatorHelpers.NodeConnection) iface.NodeClient {
+	grpcClient := grpcApi.NewNodeClient(validatorConn)
+	if features.Get().EnableBeaconRESTApi {
+		return beaconApi.NewNodeClientWithFallback(validatorConn.GetRestHandler(), grpcClient)
+	}
+	return grpcClient
+}
 
 func NewChainClient(validatorConn *validatorHelpers.NodeConnection) iface.ChainClient {
 	grpcClient := grpcApi.NewGrpcChainClient(validatorConn)
@@ -19,7 +26,14 @@ func NewChainClient(validatorConn *validatorHelpers.NodeConnection) iface.ChainC
 
 func NewPrysmChainClient(validatorConn *validatorHelpers.NodeConnection) iface.PrysmChainClient {
 	if features.Get().EnableBeaconRESTApi {
-		return beaconApi.NewPrysmChainClient(validatorConn.GetRestHandler(), nodeClientFactory.NewNodeClient(validatorConn))
+		return beaconApi.NewPrysmChainClient(validatorConn.GetRestHandler(), NewNodeClient(validatorConn))
 	}
 	return grpcApi.NewGrpcPrysmChainClient(validatorConn)
+}
+
+func NewValidatorClient(validatorConn *validatorHelpers.NodeConnection, opts ...iface.Option) iface.ValidatorClient {
+	if features.Get().EnableBeaconRESTApi {
+		return beaconApi.NewBeaconApiValidatorClient(validatorConn.GetRestConnectionProvider(), opts...)
+	}
+	return grpcApi.NewGrpcValidatorClient(validatorConn, opts...)
 }
