--- validator/client/validator-client-factory/validator_client_factory.go
@@ -1,19 +1,0 @@
-package validator_client_factory
-
-import (
-	"github.com/OffchainLabs/prysm/v7/config/features"
-	beaconApi "github.com/OffchainLabs/prysm/v7/validator/client/beacon-api"
-	grpcApi "github.com/OffchainLabs/prysm/v7/validator/client/grpc-api"
-	"github.com/OffchainLabs/prysm/v7/validator/client/iface"
-	validatorHelpers "github.com/OffchainLabs/prysm/v7/validator/helpers"
-)
-
-func NewValidatorClient(
-	validatorConn *validatorHelpers.NodeConnection,
-	opts ...iface.Option,
-) iface.ValidatorClient {
-	if features.Get().EnableBeaconRESTApi {
-		return beaconApi.NewBeaconApiValidatorClient(validatorConn.GetRestConnectionProvider(), opts...)
-	}
-	return grpcApi.NewGrpcValidatorClient(validatorConn, opts...)
-}
