+++ validator/rpc/BUILD.bazel
@@ -54,10 +54,7 @@ go_library(
         "//validator/accounts/petnames:go_default_library",
         "//validator/accounts/wallet:go_default_library",
         "//validator/client:go_default_library",
-        "//validator/client/beacon-chain-client-factory:go_default_library",
         "//validator/client/iface:go_default_library",
-        "//validator/client/node-client-factory:go_default_library",
-        "//validator/client/validator-client-factory:go_default_library",
         "//validator/db:go_default_library",
         "//validator/helpers:go_default_library",
         "//validator/keymanager:go_default_library",
